// React/app/GerAdv_TS/Apenso/Interfaces/interface.Apenso.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IApenso {
// 202501251
    id: number;
 
	processo: number,
	apensox: string,
	acao: string,
	dtdist: string,
	obs: string,
	valorcausa: number,
}

export interface IApensoFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IApensoIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}